requirements.txt
streamlit
pandas
numpy
pydantic
PyYAML
python-docx
reportlab
pytest
openpyxl
plotly
